requirements.txt
# Основные зависимости
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
pyyaml>=6.0

# Логирование и утилиты
python-dateutil>=2.8.0
pytz>=2023.3

# Обработка данных
scipy>=1.10.0
scikit-learn>=1.3.0

# Тестирование
pytest>=7.0.0
pytest-cov>=4.0.0

# Разработка
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
